+++ coral-component-table/src/tests/test.Table.js
@@ -3099,7 +3099,7 @@ describe('Table', function () {
         helpers.next(() => {
           expect(col.sortableDirection).to.equal(Table.Column.sortableDirection.DEFAULT);
           expect(headerCell.getAttribute('sortabledirection')).to.equal(col.sortableDirection);
-          expect(headerCell._elements.content.getAttribute('aria-sort')).to.equal('none');
+          expect(headerCell.getAttribute('aria-sort')).to.equal('none');
           done();
         });
       });
@@ -3116,7 +3116,7 @@ describe('Table', function () {
           expect(eventSpy.args[0][0].detail.column).to.equal(col);
           expect(col.sortableDirection).to.equal(Table.Column.sortableDirection.ASCENDING);
           expect(headerCell.getAttribute('sortabledirection')).to.equal(col.sortableDirection);
-          expect(headerCell._elements.content.getAttribute('aria-sort')).to.equal(col.sortableDirection);
+          expect(headerCell.getAttribute('aria-sort')).to.equal(col.sortableDirection);
           done();
         });
       });
@@ -3134,7 +3134,7 @@ describe('Table', function () {
           expect(eventSpy.args[0][0].detail.column).to.equal(col);
           expect(col.sortableDirection).to.equal(Table.Column.sortableDirection.DESCENDING);
           expect(headerCell.getAttribute('sortabledirection')).to.equal(col.sortableDirection);
-          expect(headerCell._elements.content.getAttribute('aria-sort')).to.equal(col.sortableDirection);
+          expect(headerCell.getAttribute('aria-sort')).to.equal(col.sortableDirection);
           done();
         });
       });
@@ -3153,7 +3153,7 @@ describe('Table', function () {
           expect(eventSpy.args[0][0].detail.column).to.equal(col);
           expect(col.sortableDirection).to.equal(Table.Column.sortableDirection.DEFAULT);
           expect(headerCell.getAttribute('sortabledirection')).to.equal(col.sortableDirection);
-          expect(headerCell._elements.content.getAttribute('aria-sort')).to.equal('none');
+          expect(headerCell.getAttribute('aria-sort')).to.equal('none');
           done();
         });
       });