+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [4.16.2](https://github.com/adobe/coral-spectrum/compare/v4.16.1...v4.16.2) (2024-05-15)
+
+
+### Bug Fixes
+
+* **Accessability:** Add button role to table header interactive element ([#347](https://github.com/adobe/coral-spectrum/issues/347)) ([a17ed32](https://github.com/adobe/coral-spectrum/commit/a17ed3203802d03d1d267eb0fbd329224e7fdb00))
+
+## [4.16.1](https://github.com/adobe/coral-spectrum/compare/v4.16.0...v4.16.1) (2024-04-23)
+
+
+### Bug Fixes
+
+* **Accessibility:** Annotation swatch selector button missing an accessible name ([#341](https://github.com/adobe/coral-spectrum/issues/341)) ([a5a4c93](https://github.com/adobe/coral-spectrum/commit/a5a4c9360970d717d2f474a5fb58209d1e15835b))
+
 # [4.16.0](https://github.com/adobe/coral-spectrum/compare/v4.15.29...v4.16.0) (2024-04-02)